Moved a few supporting methods for reformatting field values out of index and into util. Per-report submissions now format fields and values in the same way as the variant detaisl section. Removed last commented-out reference to the mock report data.
